broke login_test into several files
